Package: update Python version support
Python 3.6 has been EOL since 2021-12-23 and Python 3.9 and Python 3.10
have been released on 2020-10-05 and 2021-10-04, respectively.

The `setup.cfg` is updated to require Python 3.7 as a minimum and trove
classifiers are added for Python 3.9 and 3.10.

The CI workflow is updated to test against the new Python versions. The
version specifiers are wrapped in quotes, because without it they are
interpreted as floats first and `3.10` would become `3.1`.